+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.1.0 - New settings & timing tweaks - 2020/06/21
+* Added tweakable timeout offsets for startup and new audio
+* Changed default startup timeout to 2 seconds, as it was breaking on larger scenes
+
 ## 1.0.0 - Initial public release - 2020/06/21
 * Updated settings format
 * Squashed some bugs